refactor GetItalianMealsForLargeGroups UseCase Test
